Move all utils to sdk
