+++ scripts/add-icon-data.js
@@ -8,8 +8,8 @@ import {
   getIconsDataString,
   titleToSlug,
   normalizeColor,
+  writeIconsData,
 } from '../sdk.mjs';
-import { writeIconsData } from './utils.js';
 
 const hexPattern = /^#?[a-f0-9]{3,8}$/i;
